requirements.txt
# Core Django packages
Django==4.2.7
python-decouple==3.8

# Database
psycopg2-binary==2.9.7  # For PostgreSQL (optional)

# Django REST Framework and Authentication
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3

# Async and Background Tasks
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# WebSockets for Real-time Features
channels==4.0.0
channels-redis==4.1.0

# Optimization Libraries
ortools==9.8.3296
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Image Processing
Pillow==10.1.0

# Development and Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0

# Monitoring and Logging
sentry-sdk==1.38.0

# Documentation
django-extensions==3.2.3

# Excel/CSV Processing (for data import/export)
openpyxl==3.1.2
xlsxwriter==3.1.9

# API Documentation
drf-spectacular==0.26.5

# Geographic/Mapping Support
django-location-field==2.1.0

# Time Zone Support
pytz==2023.3

# Environment Variables
python-dotenv==1.0.0

# HTTP Client for API integrations
requests==2.31.0

# Data Validation
marshmallow==3.20.1

# Task Scheduling
django-crontab==0.7.1

# File Storage (for cloud deployments)
django-storages==1.14.2
boto3==1.29.7
